requirements.txt
Django>=1.11.19
djangorestframework==3.9.0
django-rest-swagger==2.2.0
django-extensions==2.1.4
jsonfield==2.0.2
git+ssh://git@github.com/bastionhost/django-admin-json-editor.git
git+ssh://git@github.com/bastionhost/django-json-schema-editor.git
newrelic
acm-sdk-python>=0.4.10
shortuuid
